server/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.2
django-allauth==65.7.0
django-redis==5.4.0
django-rest-swagger==2.2.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework-simplejwt==3.2.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
html-testRunner==1.2.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==25.0
prompt_toolkit==3.0.51
# psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ratelimit==2.2.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==79.0.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
